LU-9680 lnet: add NLM_F_DUMP_FILTERED support
In addition to different API levels for the netlink packets we
can also filter the data sent back when user land sends the
NLM_F_DUMP_FILTERED. Support this across the various netlink
dumpit functions.

This work is needed for the proper support for lnetctl export
command. Update the export to work with the Netlink API. This
results in proper IPv6 support for the export command.

Test-Parameters: trivial testlist=sanity-lnet
Change-Id: I0e8993b1f9a08199f282965601781aa6fd0e4844
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/53004
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Frank Sehr <fsehr@whamcloud.com>
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>