LU-17533 llite: call merge attr on all writes
Because DIO writes do not update the inode size during the
write, when a file is closed and the LSOM update is sent,
the file size provided by the client is incorrect.

DIO writes don't cause consistency problems because ls and
other things which check the file size will get the correct
size and update the inode size then, but that just means
this issue isn't fatal - DIO should still update the inode
size.

This is best done by calling ll_merge_attr on all writes at
the end of the write, rather than just for async writes in
vvp_io_write_commit.

Signed-off-by: Patrick Farrell <patrick.farrell@oracle.com>
Change-Id: I856b319254ad7093e69e41613120e06b71f656cc
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/54016
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Qian Yingjin <qian@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>