LU-17638 util: remove newer lnetctl export handling
On the current maloo VMs lnetctl export ends up segfaulting. For
now go back to the original code until we figure out what is
different on this setup and yet it works elsewhere. The reason
for a partial reveret is other important works are ready to land
that would be delayed by a full revert.

Fixes: d3ef8f6993 ("LU-9680 lnet: add NLM_F_DUMP_FILTERED support")
Test-Parameters: trivial testlist=sanity-lnet
Signed-off-by: James Simmons <jsimmons@infradead.org>
Change-Id: Ibd3437ee619cde9667d049455d641a602ea50174
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/54436
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Frank Sehr <fsehr@whamcloud.com>
Reviewed-by: Chris Horn <chris.horn@hpe.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>