LU-17431 nodemap: allow add/del of a dynamic nodemap
Adding a dynamic nodemap requires specifying the '-d' option to
'lctl nodemap_add'. This in turn sends the appropriate ioctl to
the MDS or OSS device.
Deleting a dynamic nodemap does not require this flag, as we can
figure out if the nodemap to be deleted is dynamic or not.
On kernel side, we need to allow handling a nodemap even if the
nodemap config file is not accessible (i.e. we are not on the MGS).
This dynamic nodemap is not stored on disk.
And we prevent modifications of static, on-disk nodemaps on non-MGS
servers. On these servers it is only possible to modify dynamic
nodemaps.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I10019ad131825c3db520689e01b4c931d90c1c89
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/54507
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Oleg Drokin <green@whamcloud.com>