LU-17431 nodemap: allow modifying properties of a dynamic nm
Modify properties of a dynamic nodemap, by sending the appropriate
ioctl to the MDS or OSS device.
On kernel side, we need to allow handling a nodemap even if the
nodemap config file is not accessible (i.e. we are not on the
MGS). This dynamic nodemap is not stored on disk.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I7253e51ecb8869c2d2ac268926c1c907005ec4a7
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/54513
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Marc Vef <mvef@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>