LU-17675 tests: sanity-flr/61a set atime_diff=1 for statx
Lower mdd.*.atime_diff=1 for this test so that the MDT inode atime
is updated during the test.  Otherwise, statx() will get only MDT
atime and not OST atime, and the test is not running long enough
for the MDT inode atime to be updated.

Test-Parameters: trivial testlist=sanity env=ONLY=61a,ONLY_REPEAT=100 clientdistro=el9.3
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I6518260d0dcb283faf4bc6a8049cd61abaf6978b
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/54647
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>