LU-17525 tests: fix sanity hash 2.15 interop
Fix test version checks for interop testing for DNE directory hash
usage in sanity with 2.15 servers.  This incorrectly was assuming
that the CRUSH2 dir hash was included in the 2.15.0 release, but it
was not backported to that branch, and only landed in 2.15.51.

Exclude UDIO interop failures, which are fixed via LU-17525.

Fixes: 1ac4b9598a ("LU-15720 dne: add crush2 hash type")
Test-Parameters: trivial testlist=sanity serverversion=2.15.4 serverdistro=el8.9 env=SANITY_EXCEPT="56 119 398"
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: If2097ebc30c7c4dbce88af7774ce3c0e8fb3cb75
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/55076
Tested-by: Maloo <maloo@whamcloud.com>
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Patrick Farrell <patrick.farrell@oracle.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>